+++ src/main/java/com/almor/course_project/repos/UsersRatingRepo.java
@@ -8,8 +8,8 @@ public interface UsersRatingRepo extends CrudRepository<UsersRatings, Long> {
 
     @Query("SELECT AVG(ur.rating) FROM UsersRatings ur " +
             "WHERE ur.id.campaignId = ?1 GROUP BY ur.id.campaignId")
-    Integer findAvgCampaignRating(int campaignId);
+    Integer findAvgCampaignRating(Long campaignId);
 
     @Query("SELECT ur.rating FROM UsersRatings ur WHERE ur.id.campaignId = ?1 AND ur.id.userId = ?2")
-    Integer findUserCampaignRating(int campaignId, int userId);
+    Integer findUserCampaignRating(Long campaignId, Long userId);
 }
